Move tests to Go built-in testing infrastructure
